Rewrite the entire library to provide a better API
... hopefully!

The new API is designed to allow chaining of settings and operations.

Also, uploading and downloading files can be done through readers and
writers. For instance, this allows you to track the progress of the
operation.

The new library comes with a new testsuite, which should cover way more
code than the previous one.

[#155972545]